notebooks/requirements.txt
flask
google
httpx
langchain
openai
Pillow
pydantic
pydantic-settings
python-dotenv
stripe